+++ src/main/java/calculator/Calculation.java
@@ -1,7 +1,8 @@
 package calculator;
 
+import java.util.function.IntBinaryOperator;
+
 public class Calculation {
-    TwoOperandCalculator twoOperandCalculator = TwoOperandCalculator.getOperatorSelection();
 
     public int calculate(StringExpression expression) {
         String[] splitedExpression = expression.getSplitedExpression();
@@ -9,10 +10,17 @@ public class Calculation {
         int result = Integer.parseInt(splitedExpression[0]);
         for (int i = 1; i < splitedExpression.length - 1; i += 2) {
             int operand = Integer.parseInt(splitedExpression[i + 1]);
-            result = twoOperandCalculator.calculateTwoOperands(result, operand, splitedExpression[i]);
+            result = calculateTwoOperands(result, operand, splitedExpression[i]);
         }
-
         return result;
+    }
+
+    public IntBinaryOperator findOperator(String operatorToFind) {
+        return Operator.getOperatorMethod(operatorToFind);
+    }
+
+    public int calculateTwoOperands(int firstOperand, int secondOperand, String operator) {
+        return findOperator(operator).applyAsInt(firstOperand, secondOperand);
     }
 
 }
