+++ src/main/java/lotto/domain/Lotto.java
@@ -1,0 +1,4 @@
+package lotto;
+
+public class Lotto {
+}
