+++ src/test/java/calculator/CalculationTest.java
@@ -1,17 +1,19 @@
 package calculator;
 
+import calculator.exception.DivisionCannotBeZeroException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class OperatorUtilsTest {
+class CalculationTest {
     @Test
     @DisplayName("덧셈 테스트")
     void 덧셈_테스트() {
 
+        StringExpression expression = new StringExpression("2 + 5");
         StringCalculator calculator = new StringCalculator();
 
-        Assertions.assertThat(calculator.calculate("2 + 5")).isEqualTo(7);
+        Assertions.assertThat(calculator.calculateExpression(expression)).isEqualTo(7);
 
     }
 
@@ -19,9 +21,10 @@ class OperatorUtilsTest {
     @DisplayName("뺄셈 테스트")
     void 뺄셈_테스트() {
 
+        StringExpression expression = new StringExpression("5 - 2");
         StringCalculator calculator = new StringCalculator();
 
-        Assertions.assertThat(calculator.calculate("5 - 2")).isEqualTo(3);
+        Assertions.assertThat(calculator.calculateExpression(expression)).isEqualTo(3);
 
     }
 
@@ -29,9 +32,10 @@ class OperatorUtilsTest {
     @DisplayName("곱셈 테스트")
     void 곱셈_테스트() {
 
+        StringExpression expression = new StringExpression("5 * 2");
         StringCalculator calculator = new StringCalculator();
 
-        Assertions.assertThat(calculator.calculate("5 * 2")).isEqualTo(10);
+        Assertions.assertThat(calculator.calculateExpression(expression)).isEqualTo(10);
 
     }
 
@@ -40,9 +44,10 @@ class OperatorUtilsTest {
     @DisplayName("나눗셈 테스트")
     void 나눗셈_테스트() {
 
+        StringExpression expression = new StringExpression("5 / 2");
         StringCalculator calculator = new StringCalculator();
 
-        Assertions.assertThat(calculator.calculate("5 / 2")).isEqualTo(2);
+        Assertions.assertThat(calculator.calculateExpression(expression)).isEqualTo(2);
 
     }
 
@@ -51,9 +56,10 @@ class OperatorUtilsTest {
     @DisplayName("복잡한 연산 테스트")
     void 복잡한_연산_테스트() {
 
+        StringExpression expression = new StringExpression("2 + 3 * 4 / 2");
         StringCalculator calculator = new StringCalculator();
 
-        Assertions.assertThat(calculator.calculate("2 + 3 * 4 / 2")).isEqualTo(10);
+        Assertions.assertThat(calculator.calculateExpression(expression)).isEqualTo(10);
 
     }
 
@@ -61,10 +67,12 @@ class OperatorUtilsTest {
     @DisplayName("잘못된 연산기호인지 확인한다")
     void 잘못된_연산기호_테스트() {
 
+        Calculation calculation = new Calculation();
+
         Assertions.assertThatExceptionOfType(IllegalArgumentException.class)
                 .isThrownBy(() ->
-                        OperatorUtils.selection("^"))
-                .withMessage("잘못된 연산기호 입니다");
+                        calculation.findOperator("^"))
+                .withMessage("^는 잘못된 연산기호 입니다");
 
     }
 
@@ -72,12 +80,13 @@ class OperatorUtilsTest {
     @DisplayName("0으로 나눌 시 예외 발생")
     void exceptionOccuredWhenDivideByZero() {
 
+        StringExpression expression = new StringExpression("0 / 0");
         StringCalculator calculator = new StringCalculator();
 
-        Assertions.assertThatExceptionOfType(ArithmeticException.class)
+        Assertions.assertThatExceptionOfType(DivisionCannotBeZeroException.class)
                 .isThrownBy(() ->
-                        calculator.calculate("5 / 0"))
-                .withMessage("0으로 나눌 수 없습니다");
+                        calculator.calculateExpression(expression))
+                .withMessage("0 으로 나눌 수 없습니다");
 
     }
